Correct Grid Size and Use Random Colors
Corrected the grid size by setting the container width and height to
the same value (800) and then calculating square width and height as
800 / size when 'choose size' button is clicked. The grid is also
squared whatever value is given (between 1 and 100). If the user enters
20, the grid will be 20 x 20 (400). The random color bit is for extra
credit. I tried to do a random rgb value, but found it was easier to
use a random hex color assignemnt.